Fix go back button to work
