+++ app/src/main/java/net/exkazuu/mimicdance/pages/lesson/judge/DuoJudgeFragment.java
@@ -2,6 +2,7 @@ package net.exkazuu.mimicdance.pages.lesson.judge;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
+import android.support.v4.app.FragmentManager;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -24,6 +25,7 @@ import java.util.List;
 
 import butterknife.Bind;
 import butterknife.ButterKnife;
+import butterknife.OnClick;
 import jp.fkmsoft.android.framework.util.FragmentUtils;
 
 public class DuoJudgeFragment extends BaseJudgeFragment {
@@ -134,5 +136,14 @@ public class DuoJudgeFragment extends BaseJudgeFragment {
         };
 
         robotExecutor.start();
+    }
+
+    @OnClick(R.id.button_lesson_editor_duo)
+    void lessonEditorClicked() {
+        FragmentManager manager = getFragmentManager();
+        if (manager == null) {
+            return;
+        }
+        manager.popBackStack();
     }
 }
