+++ app/src/main/res/layout/fragment_judge.xml
@@ -85,7 +85,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_weight="1"
-        android:onClick="startCodingActivity"
         android:text="@string/try_again"/>
     </LinearLayout>
   </LinearLayout>
